# SSE2 acceleration for shuffling larger types