--- a/blosc/blosc.c
+++ b/blosc/blosc.c
@@ -871,19 +871,10 @@ static int32_t compute_blocksize(struct blosc_context* context, int32_t clevel, 
     }
   }
   else if (nbytes > (16 * 16))  {
-      /* align to typesize to make use of vectorized shuffles */
-      if (typesize == 2) {
-          blocksize -= blocksize % (16 * 2);
-      }
-      else if (typesize == 4) {
-          blocksize -= blocksize % (16 * 4);
-      }
-      else if (typesize == 8) {
-          blocksize -= blocksize % (16 * 8);
-      }
-      else if (typesize == 16) {
-          blocksize -= blocksize % (16 * 16);
-      }
+    /* align to typesize to make use of vectorized shuffles */
+		if (typesize == 2 || typesize == 4 || typesize == 8 || typesize == 16 || typesize > 16) {
+			blocksize -= blocksize % (16 * typesize);
+    }
   }
 
   /* Check that blocksize is not too large */
@@ -891,9 +882,9 @@ static int32_t compute_blocksize(struct blosc_context* context, int32_t clevel, 
     blocksize = nbytes;
   }
 
-  /* blocksize must be a multiple of the typesize */
+  /* blocksize must be a multiple of the typesize and a multiple of the vector size */
   if (blocksize > typesize) {
-    blocksize = blocksize / typesize * typesize;
+		blocksize -= blocksize % (16 * typesize);
   }
 
   /* blocksize must not exceed (64 KB * typesize) in order to allow