--- a/blosc/shuffle.c
+++ b/blosc/shuffle.c
@@ -470,24 +470,22 @@ void shuffle(size_t bytesoftype, size_t blocksize,
 static void
 unshuffle2(uint8_t* dest, const uint8_t* orig, size_t size)
 {
-	size_t i, k;
-	size_t neblock, numof16belem;
+	size_t i, j;
+	const size_t numof16belem = size / (2 * VECTOR_SIZE);
 	__m128i xmm1[2], xmm2[2];
 
-	neblock = size / 2;
-	numof16belem = neblock / 16;
-	for (i = 0, k = 0; i < numof16belem; i++, k += 2) {
-		/* Load the first 32 bytes in 2 XMM registrers */
+	for (i = 0, j = 0; i < numof16belem; i++, j += 2) {
+		/* Load the first 32 bytes in 2 XMM registers */
 		xmm1[0] = ((__m128i *)orig)[0 * numof16belem + i];
 		xmm1[1] = ((__m128i *)orig)[1 * numof16belem + i];
-		/* Shuffle bytes */
+		/* Unshuffle bytes */
 		/* Compute the low 32 bytes */
 		xmm2[0] = _mm_unpacklo_epi8(xmm1[0], xmm1[1]);
 		/* Compute the hi 32 bytes */
 		xmm2[1] = _mm_unpackhi_epi8(xmm1[0], xmm1[1]);
 		/* Store the result vectors in proper order */
-		((__m128i *)dest)[k + 0] = xmm2[0];
-		((__m128i *)dest)[k + 1] = xmm2[1];
+		((__m128i *)dest)[j + 0] = xmm2[0];
+		((__m128i *)dest)[j + 1] = xmm2[1];
 	}
 }
 
@@ -496,36 +494,36 @@ unshuffle2(uint8_t* dest, const uint8_t* orig, size_t size)
 static void
 unshuffle4(uint8_t* dest, const uint8_t* orig, size_t size)
 {
-	size_t i, j, k;
-	size_t neblock, numof16belem;
+	size_t i, j;
+	const size_t numof16belem = size / (4 * VECTOR_SIZE);
 	__m128i xmm0[4], xmm1[4];
 
-	neblock = size / 4;
-	numof16belem = neblock / 16;
-	for (i = 0, k = 0; i < numof16belem; i++, k += 4) {
-		/* Load the first 64 bytes in 4 XMM registrers */
-		for (j = 0; j < 4; j++) {
-			xmm0[j] = ((__m128i *)orig)[j*numof16belem + i];
-		}
-		/* Shuffle bytes */
-		for (j = 0; j < 2; j++) {
-			/* Compute the low 32 bytes */
-			xmm1[j] = _mm_unpacklo_epi8(xmm0[j * 2], xmm0[j * 2 + 1]);
-			/* Compute the hi 32 bytes */
-			xmm1[2 + j] = _mm_unpackhi_epi8(xmm0[j * 2], xmm0[j * 2 + 1]);
-		}
-		/* Shuffle 2-byte words */
-		for (j = 0; j < 2; j++) {
-			/* Compute the low 32 bytes */
-			xmm0[j] = _mm_unpacklo_epi16(xmm1[j * 2], xmm1[j * 2 + 1]);
-			/* Compute the hi 32 bytes */
-			xmm0[2 + j] = _mm_unpackhi_epi16(xmm1[j * 2], xmm1[j * 2 + 1]);
-		}
+	for (i = 0, j = 0; i < numof16belem; i++, j += 4) {
+		/* Load the first 64 bytes in 4 XMM registers */
+		xmm0[0] = ((__m128i *)orig)[0*numof16belem + i];
+		xmm0[1] = ((__m128i *)orig)[1*numof16belem + i];
+		xmm0[2] = ((__m128i *)orig)[2*numof16belem + i];
+		xmm0[3] = ((__m128i *)orig)[3*numof16belem + i];
+
+		/* Unshuffle bytes */
+		xmm1[0] = _mm_unpacklo_epi8(xmm0[0], xmm0[1]);
+		xmm1[2] = _mm_unpackhi_epi8(xmm0[0], xmm0[1]);
+
+		xmm1[1] = _mm_unpacklo_epi8(xmm0[2], xmm0[3]);
+		xmm1[3] = _mm_unpackhi_epi8(xmm0[2], xmm0[3]);
+
+		/* Unshuffle 2-byte words */
+		xmm0[0] = _mm_unpacklo_epi16(xmm1[0], xmm1[1]);
+		xmm0[2] = _mm_unpackhi_epi16(xmm1[0], xmm1[1]);
+
+		xmm0[1] = _mm_unpacklo_epi16(xmm1[2], xmm1[3]);
+		xmm0[3] = _mm_unpackhi_epi16(xmm1[2], xmm1[3]);
+
 		/* Store the result vectors in proper order */
-		((__m128i *)dest)[k + 0] = xmm0[0];
-		((__m128i *)dest)[k + 1] = xmm0[2];
-		((__m128i *)dest)[k + 2] = xmm0[1];
-		((__m128i *)dest)[k + 3] = xmm0[3];
+		((__m128i *)dest)[j + 0] = xmm0[0];
+		((__m128i *)dest)[j + 1] = xmm0[2];
+		((__m128i *)dest)[j + 2] = xmm0[1];
+		((__m128i *)dest)[j + 3] = xmm0[3];
 	}
 }
 
@@ -534,47 +532,69 @@ unshuffle4(uint8_t* dest, const uint8_t* orig, size_t size)
 static void
 unshuffle8(uint8_t* dest, const uint8_t* orig, size_t size)
 {
-	size_t i, j, k;
-	size_t neblock, numof16belem;
+	size_t i, j;
+	const size_t numof16belem = size / (8 * VECTOR_SIZE);
 	__m128i xmm0[8], xmm1[8];
 
-	neblock = size / 8;
-	numof16belem = neblock / 16;
-	for (i = 0, k = 0; i < numof16belem; i++, k += 8) {
-		/* Load the first 64 bytes in 8 XMM registrers */
-		for (j = 0; j < 8; j++) {
-			xmm0[j] = ((__m128i *)orig)[j*numof16belem + i];
-		}
-		/* Shuffle bytes */
-		for (j = 0; j < 4; j++) {
-			/* Compute the low 32 bytes */
-			xmm1[j] = _mm_unpacklo_epi8(xmm0[j * 2], xmm0[j * 2 + 1]);
-			/* Compute the hi 32 bytes */
-			xmm1[4 + j] = _mm_unpackhi_epi8(xmm0[j * 2], xmm0[j * 2 + 1]);
-		}
-		/* Shuffle 2-byte words */
-		for (j = 0; j < 4; j++) {
-			/* Compute the low 32 bytes */
-			xmm0[j] = _mm_unpacklo_epi16(xmm1[j * 2], xmm1[j * 2 + 1]);
-			/* Compute the hi 32 bytes */
-			xmm0[4 + j] = _mm_unpackhi_epi16(xmm1[j * 2], xmm1[j * 2 + 1]);
-		}
-		/* Shuffle 4-byte dwords */
-		for (j = 0; j < 4; j++) {
-			/* Compute the low 32 bytes */
-			xmm1[j] = _mm_unpacklo_epi32(xmm0[j * 2], xmm0[j * 2 + 1]);
-			/* Compute the hi 32 bytes */
-			xmm1[4 + j] = _mm_unpackhi_epi32(xmm0[j * 2], xmm0[j * 2 + 1]);
-		}
+	for (i = 0, j = 0; i < numof16belem; i++, j += 8) {
+		/* Load the first 64 bytes in 8 XMM registers */
+		xmm0[0] = ((__m128i *)orig)[0*numof16belem + i];
+		xmm0[1] = ((__m128i *)orig)[1*numof16belem + i];
+		xmm0[2] = ((__m128i *)orig)[2*numof16belem + i];
+		xmm0[3] = ((__m128i *)orig)[3*numof16belem + i];
+		xmm0[4] = ((__m128i *)orig)[4*numof16belem + i];
+		xmm0[5] = ((__m128i *)orig)[5*numof16belem + i];
+		xmm0[6] = ((__m128i *)orig)[6*numof16belem + i];
+		xmm0[7] = ((__m128i *)orig)[7*numof16belem + i];
+
+		/* Unshuffle bytes */
+		xmm1[0] = _mm_unpacklo_epi8(xmm0[0], xmm0[1]);
+		xmm1[4] = _mm_unpackhi_epi8(xmm0[0], xmm0[1]);
+
+		xmm1[1] = _mm_unpacklo_epi8(xmm0[2], xmm0[3]);
+		xmm1[5] = _mm_unpackhi_epi8(xmm0[2], xmm0[3]);
+
+		xmm1[2] = _mm_unpacklo_epi8(xmm0[4], xmm0[5]);
+		xmm1[6] = _mm_unpackhi_epi8(xmm0[4], xmm0[5]);
+
+		xmm1[3] = _mm_unpacklo_epi8(xmm0[6], xmm0[7]);
+		xmm1[7] = _mm_unpackhi_epi8(xmm0[6], xmm0[7]);
+
+		/* Unshuffle 2-byte words */
+		xmm0[0] = _mm_unpacklo_epi16(xmm1[0], xmm1[1]);
+		xmm0[4] = _mm_unpackhi_epi16(xmm1[0], xmm1[1]);
+
+		xmm0[1] = _mm_unpacklo_epi16(xmm1[2], xmm1[3]);
+		xmm0[5] = _mm_unpackhi_epi16(xmm1[2], xmm1[3]);
+
+		xmm0[2] = _mm_unpacklo_epi16(xmm1[4], xmm1[5]);
+		xmm0[6] = _mm_unpackhi_epi16(xmm1[4], xmm1[5]);
+
+		xmm0[3] = _mm_unpacklo_epi16(xmm1[6], xmm1[7]);
+		xmm0[7] = _mm_unpackhi_epi16(xmm1[6], xmm1[7]);
+
+		/* Unshuffle 4-byte dwords */
+		xmm1[0] = _mm_unpacklo_epi32(xmm0[0], xmm0[1]);
+		xmm1[4] = _mm_unpackhi_epi32(xmm0[0], xmm0[1]);
+
+		xmm1[1] = _mm_unpacklo_epi32(xmm0[2], xmm0[3]);
+		xmm1[5] = _mm_unpackhi_epi32(xmm0[2], xmm0[3]);
+
+		xmm1[2] = _mm_unpacklo_epi32(xmm0[4], xmm0[5]);
+		xmm1[6] = _mm_unpackhi_epi32(xmm0[4], xmm0[5]);
+
+		xmm1[3] = _mm_unpacklo_epi32(xmm0[6], xmm0[7]);
+		xmm1[7] = _mm_unpackhi_epi32(xmm0[6], xmm0[7]);
+
 		/* Store the result vectors in proper order */
-		((__m128i *)dest)[k + 0] = xmm1[0];
-		((__m128i *)dest)[k + 1] = xmm1[4];
-		((__m128i *)dest)[k + 2] = xmm1[2];
-		((__m128i *)dest)[k + 3] = xmm1[6];
-		((__m128i *)dest)[k + 4] = xmm1[1];
-		((__m128i *)dest)[k + 5] = xmm1[5];
-		((__m128i *)dest)[k + 6] = xmm1[3];
-		((__m128i *)dest)[k + 7] = xmm1[7];
+		((__m128i *)dest)[j + 0] = xmm1[0];
+		((__m128i *)dest)[j + 1] = xmm1[4];
+		((__m128i *)dest)[j + 2] = xmm1[2];
+		((__m128i *)dest)[j + 3] = xmm1[6];
+		((__m128i *)dest)[j + 4] = xmm1[1];
+		((__m128i *)dest)[j + 5] = xmm1[5];
+		((__m128i *)dest)[j + 6] = xmm1[3];
+		((__m128i *)dest)[j + 7] = xmm1[7];
 	}
 }
 
@@ -583,62 +603,146 @@ unshuffle8(uint8_t* dest, const uint8_t* orig, size_t size)
 static void
 unshuffle16(uint8_t* dest, const uint8_t* orig, size_t size)
 {
-	size_t i, j, k;
-	size_t neblock, numof16belem;
+	size_t i, j;
+	const size_t numof16belem = size / (16 * VECTOR_SIZE);
 	__m128i xmm1[16], xmm2[16];
 
-	neblock = size / 16;
-	numof16belem = neblock / 16;
-	for (i = 0, k = 0; i < numof16belem; i++, k += 16) {
-		/* Load the first 128 bytes in 16 XMM registrers */
-		for (j = 0; j < 16; j++) {
-			xmm1[j] = ((__m128i *)orig)[j*numof16belem + i];
-		}
-		/* Shuffle bytes */
-		for (j = 0; j < 8; j++) {
-			/* Compute the low 32 bytes */
-			xmm2[j] = _mm_unpacklo_epi8(xmm1[j * 2], xmm1[j * 2 + 1]);
-			/* Compute the hi 32 bytes */
-			xmm2[8 + j] = _mm_unpackhi_epi8(xmm1[j * 2], xmm1[j * 2 + 1]);
-		}
-		/* Shuffle 2-byte words */
-		for (j = 0; j < 8; j++) {
-			/* Compute the low 32 bytes */
-			xmm1[j] = _mm_unpacklo_epi16(xmm2[j * 2], xmm2[j * 2 + 1]);
-			/* Compute the hi 32 bytes */
-			xmm1[8 + j] = _mm_unpackhi_epi16(xmm2[j * 2], xmm2[j * 2 + 1]);
-		}
-		/* Shuffle 4-byte dwords */
-		for (j = 0; j < 8; j++) {
-			/* Compute the low 32 bytes */
-			xmm2[j] = _mm_unpacklo_epi32(xmm1[j * 2], xmm1[j * 2 + 1]);
-			/* Compute the hi 32 bytes */
-			xmm2[8 + j] = _mm_unpackhi_epi32(xmm1[j * 2], xmm1[j * 2 + 1]);
-		}
-		/* Shuffle 8-byte qwords */
-		for (j = 0; j < 8; j++) {
-			/* Compute the low 32 bytes */
-			xmm1[j] = _mm_unpacklo_epi64(xmm2[j * 2], xmm2[j * 2 + 1]);
-			/* Compute the hi 32 bytes */
-			xmm1[8 + j] = _mm_unpackhi_epi64(xmm2[j * 2], xmm2[j * 2 + 1]);
-		}
+	for (i = 0, j = 0; i < numof16belem; i++, j += 16) {
+		/* Load the first 128 bytes in 16 XMM registers */
+		xmm1[0] = ((__m128i *)orig)[0*numof16belem + i];
+		xmm1[1] = ((__m128i *)orig)[1*numof16belem + i];
+		xmm1[2] = ((__m128i *)orig)[2*numof16belem + i];
+		xmm1[3] = ((__m128i *)orig)[3*numof16belem + i];
+		xmm1[4] = ((__m128i *)orig)[4*numof16belem + i];
+		xmm1[5] = ((__m128i *)orig)[5*numof16belem + i];
+		xmm1[6] = ((__m128i *)orig)[6*numof16belem + i];
+		xmm1[7] = ((__m128i *)orig)[7*numof16belem + i];
+		xmm1[8] = ((__m128i *)orig)[8*numof16belem + i];
+		xmm1[9] = ((__m128i *)orig)[9*numof16belem + i];
+		xmm1[10] = ((__m128i *)orig)[10*numof16belem + i];
+		xmm1[11] = ((__m128i *)orig)[11*numof16belem + i];
+		xmm1[12] = ((__m128i *)orig)[12*numof16belem + i];
+		xmm1[13] = ((__m128i *)orig)[13*numof16belem + i];
+		xmm1[14] = ((__m128i *)orig)[14*numof16belem + i];
+		xmm1[15] = ((__m128i *)orig)[15*numof16belem + i];
+
+		/* Unshuffle bytes */
+		xmm2[0] = _mm_unpacklo_epi8(xmm1[0], xmm1[1]);
+		xmm2[8] = _mm_unpackhi_epi8(xmm1[0], xmm1[1]);
+
+		xmm2[1] = _mm_unpacklo_epi8(xmm1[2], xmm1[3]);
+		xmm2[9] = _mm_unpackhi_epi8(xmm1[2], xmm1[3]);
+
+		xmm2[2] = _mm_unpacklo_epi8(xmm1[4], xmm1[5]);
+		xmm2[10] = _mm_unpackhi_epi8(xmm1[4], xmm1[5]);
+
+		xmm2[3] = _mm_unpacklo_epi8(xmm1[6], xmm1[7]);
+		xmm2[11] = _mm_unpackhi_epi8(xmm1[6], xmm1[7]);
+
+		xmm2[4] = _mm_unpacklo_epi8(xmm1[8], xmm1[9]);
+		xmm2[12] = _mm_unpackhi_epi8(xmm1[8], xmm1[9]);
+
+		xmm2[5] = _mm_unpacklo_epi8(xmm1[10], xmm1[11]);
+		xmm2[13] = _mm_unpackhi_epi8(xmm1[10], xmm1[11]);
+
+		xmm2[6] = _mm_unpacklo_epi8(xmm1[12], xmm1[13]);
+		xmm2[14] = _mm_unpackhi_epi8(xmm1[12], xmm1[13]);
+
+		xmm2[7] = _mm_unpacklo_epi8(xmm1[14], xmm1[15]);
+		xmm2[15] = _mm_unpackhi_epi8(xmm1[14], xmm1[15]);
+
+		/* Unshuffle 2-byte words */
+		xmm1[0] = _mm_unpacklo_epi16(xmm2[0], xmm2[1]);
+		xmm1[8] = _mm_unpackhi_epi16(xmm2[0], xmm2[1]);
+
+		xmm1[1] = _mm_unpacklo_epi16(xmm2[2], xmm2[3]);
+		xmm1[9] = _mm_unpackhi_epi16(xmm2[2], xmm2[3]);
+
+		xmm1[2] = _mm_unpacklo_epi16(xmm2[4], xmm2[5]);
+		xmm1[10] = _mm_unpackhi_epi16(xmm2[4], xmm2[5]);
+
+		xmm1[3] = _mm_unpacklo_epi16(xmm2[6], xmm2[7]);
+		xmm1[11] = _mm_unpackhi_epi16(xmm2[6], xmm2[7]);
+
+		xmm1[4] = _mm_unpacklo_epi16(xmm2[8], xmm2[9]);
+		xmm1[12] = _mm_unpackhi_epi16(xmm2[8], xmm2[9]);
+
+		xmm1[5] = _mm_unpacklo_epi16(xmm2[10], xmm2[11]);
+		xmm1[13] = _mm_unpackhi_epi16(xmm2[10], xmm2[11]);
+
+		xmm1[6] = _mm_unpacklo_epi16(xmm2[12], xmm2[13]);
+		xmm1[14] = _mm_unpackhi_epi16(xmm2[12], xmm2[13]);
+
+		xmm1[7] = _mm_unpacklo_epi16(xmm2[14], xmm2[15]);
+		xmm1[15] = _mm_unpackhi_epi16(xmm2[14], xmm2[15]);
+
+		/* Unshuffle 4-byte dwords */
+		xmm2[0] = _mm_unpacklo_epi32(xmm1[0], xmm1[1]);
+		xmm2[8] = _mm_unpackhi_epi32(xmm1[0], xmm1[1]);
+
+		xmm2[1] = _mm_unpacklo_epi32(xmm1[2], xmm1[3]);
+		xmm2[9] = _mm_unpackhi_epi32(xmm1[2], xmm1[3]);
+
+		xmm2[2] = _mm_unpacklo_epi32(xmm1[4], xmm1[5]);
+		xmm2[10] = _mm_unpackhi_epi32(xmm1[4], xmm1[5]);
+
+		xmm2[3] = _mm_unpacklo_epi32(xmm1[6], xmm1[7]);
+		xmm2[11] = _mm_unpackhi_epi32(xmm1[6], xmm1[7]);
+
+		xmm2[4] = _mm_unpacklo_epi32(xmm1[8], xmm1[9]);
+		xmm2[12] = _mm_unpackhi_epi32(xmm1[8], xmm1[9]);
+
+		xmm2[5] = _mm_unpacklo_epi32(xmm1[10], xmm1[11]);
+		xmm2[13] = _mm_unpackhi_epi32(xmm1[10], xmm1[11]);
+
+		xmm2[6] = _mm_unpacklo_epi32(xmm1[12], xmm1[13]);
+		xmm2[14] = _mm_unpackhi_epi32(xmm1[12], xmm1[13]);
+
+		xmm2[7] = _mm_unpacklo_epi32(xmm1[14], xmm1[15]);
+		xmm2[15] = _mm_unpackhi_epi32(xmm1[14], xmm1[15]);
+
+		/* Unshuffle 8-byte qwords */
+		xmm1[0] = _mm_unpacklo_epi64(xmm2[0], xmm2[1]);
+		xmm1[8] = _mm_unpackhi_epi64(xmm2[0], xmm2[1]);
+
+		xmm1[1] = _mm_unpacklo_epi64(xmm2[2], xmm2[3]);
+		xmm1[9] = _mm_unpackhi_epi64(xmm2[2], xmm2[3]);
+
+		xmm1[2] = _mm_unpacklo_epi64(xmm2[4], xmm2[5]);
+		xmm1[10] = _mm_unpackhi_epi64(xmm2[4], xmm2[5]);
+
+		xmm1[3] = _mm_unpacklo_epi64(xmm2[6], xmm2[7]);
+		xmm1[11] = _mm_unpackhi_epi64(xmm2[6], xmm2[7]);
+
+		xmm1[4] = _mm_unpacklo_epi64(xmm2[8], xmm2[9]);
+		xmm1[12] = _mm_unpackhi_epi64(xmm2[8], xmm2[9]);
+
+		xmm1[5] = _mm_unpacklo_epi64(xmm2[10], xmm2[11]);
+		xmm1[13] = _mm_unpackhi_epi64(xmm2[10], xmm2[11]);
+
+		xmm1[6] = _mm_unpacklo_epi64(xmm2[12], xmm2[13]);
+		xmm1[14] = _mm_unpackhi_epi64(xmm2[12], xmm2[13]);
+
+		xmm1[7] = _mm_unpacklo_epi64(xmm2[14], xmm2[15]);
+		xmm1[15] = _mm_unpackhi_epi64(xmm2[14], xmm2[15]);
+
 		/* Store the result vectors in proper order */
-		((__m128i *)dest)[k + 0] = xmm1[0];
-		((__m128i *)dest)[k + 1] = xmm1[8];
-		((__m128i *)dest)[k + 2] = xmm1[4];
-		((__m128i *)dest)[k + 3] = xmm1[12];
-		((__m128i *)dest)[k + 4] = xmm1[2];
-		((__m128i *)dest)[k + 5] = xmm1[10];
-		((__m128i *)dest)[k + 6] = xmm1[6];
-		((__m128i *)dest)[k + 7] = xmm1[14];
-		((__m128i *)dest)[k + 8] = xmm1[1];
-		((__m128i *)dest)[k + 9] = xmm1[9];
-		((__m128i *)dest)[k + 10] = xmm1[5];
-		((__m128i *)dest)[k + 11] = xmm1[13];
-		((__m128i *)dest)[k + 12] = xmm1[3];
-		((__m128i *)dest)[k + 13] = xmm1[11];
-		((__m128i *)dest)[k + 14] = xmm1[7];
-		((__m128i *)dest)[k + 15] = xmm1[15];
+		((__m128i *)dest)[j + 0] = xmm1[0];
+		((__m128i *)dest)[j + 1] = xmm1[8];
+		((__m128i *)dest)[j + 2] = xmm1[4];
+		((__m128i *)dest)[j + 3] = xmm1[12];
+		((__m128i *)dest)[j + 4] = xmm1[2];
+		((__m128i *)dest)[j + 5] = xmm1[10];
+		((__m128i *)dest)[j + 6] = xmm1[6];
+		((__m128i *)dest)[j + 7] = xmm1[14];
+		((__m128i *)dest)[j + 8] = xmm1[1];
+		((__m128i *)dest)[j + 9] = xmm1[9];
+		((__m128i *)dest)[j + 10] = xmm1[5];
+		((__m128i *)dest)[j + 11] = xmm1[13];
+		((__m128i *)dest)[j + 12] = xmm1[3];
+		((__m128i *)dest)[j + 13] = xmm1[11];
+		((__m128i *)dest)[j + 14] = xmm1[7];
+		((__m128i *)dest)[j + 15] = xmm1[15];
 	}
 }