I've implemented SSE2-accelerated shuffle/unshuffle functions which can be used when the type size is larger than the SSE2 vector type (i.e., >16 bytes).

I've also hand-unrolled all of the small inner loops within the existing shuffle/unshuffle routines (for 2/4/8/16-byte types). This could have been done by some compilers already (in theory, at least), but I ran a baseline benchmark, then benched again with the change and the performance improved noticeably. I've included the results below for comparison.

Baseline:

```
Blosc version: 1.5.3.dev ($Date:: 2014-12-30 #$)
List of supported compressors in this build: blosclz,lz4,lz4hc,zlib
Supported compression libraries:
  BloscLZ: 1.0.3
  LZ4: 1.5.0
  Snappy: unknown
  Zlib: 1.2.8
Using compressor: blosclz
Running suite: suite


--> 1, 536870912, 16, 19, blosclz, 1
********************** Run info ******************************
Blosc version: 1.5.3.dev ($Date:: 2014-12-30 #$)
Number of threads: 1
Using synthetic data with 19 significant bits (out of 32)
Working set: 512.0 MB
Dataset size: 536870912 bytes
Type size: 16 bytes
Element Count: 33554432
********************** Running benchmarks *********************
memcpy(write):		 47219.8 us, 10842.9 MB/s
memcpy(read):		 88055.2 us, 5814.5 MB/s
Compression level: 0
comp(write):	 58532.1 us, 8747.3 MB/s	  Final bytes: 536870928  Ratio: 1.00
decomp(read):	 50988.6 us, 10041.5 MB/s	  OK
Compression level: 1
comp(write):	 214390.2 us, 2388.2 MB/s	  Final bytes: 77791248  Ratio: 6.90
decomp(read):	 108376.0 us, 4724.3 MB/s	  OK
Compression level: 2
comp(write):	 230685.9 us, 2219.5 MB/s	  Final bytes: 45645840  Ratio: 11.76
decomp(read):	 118848.3 us, 4308.0 MB/s	  OK
Compression level: 3
comp(write):	 236344.4 us, 2166.3 MB/s	  Final bytes: 45645840  Ratio: 11.76
decomp(read):	 118475.5 us, 4321.6 MB/s	  OK
Compression level: 4
comp(write):	 270834.0 us, 1890.5 MB/s	  Final bytes: 40583184  Ratio: 13.23
decomp(read):	 119848.0 us, 4272.1 MB/s	  OK
Compression level: 5
comp(write):	 297610.6 us, 1720.4 MB/s	  Final bytes: 40583184  Ratio: 13.23
decomp(read):	 119642.9 us, 4279.4 MB/s	  OK
Compression level: 6
comp(write):	 308840.5 us, 1657.8 MB/s	  Final bytes: 23207952  Ratio: 23.13
decomp(read):	 142526.7 us, 3592.3 MB/s	  OK
Compression level: 7
comp(write):	 282981.9 us, 1809.3 MB/s	  Final bytes: 19865616  Ratio: 27.03
decomp(read):	 148218.5 us, 3454.4 MB/s	  OK
Compression level: 8
comp(write):	 281059.1 us, 1821.7 MB/s	  Final bytes: 19865616  Ratio: 27.03
decomp(read):	 148182.3 us, 3455.2 MB/s	  OK
Compression level: 9
comp(write):	 282498.8 us, 1812.4 MB/s	  Final bytes: 16814096  Ratio: 31.93
decomp(read):	 151238.2 us, 3385.4 MB/s	  OK

Round-trip compr/decompr on 15.0 GB
Elapsed time:	   12.6 s, 2678.3 MB/s
```

After applying this PR:

```
Blosc version: 1.5.3.dev ($Date:: 2014-12-30 #$)
List of supported compressors in this build: blosclz,lz4,lz4hc,zlib
Supported compression libraries:
  BloscLZ: 1.0.3
  LZ4: 1.5.0
  Snappy: unknown
  Zlib: 1.2.8
Using compressor: blosclz
Running suite: suite


--> 1, 536870912, 16, 19, blosclz, 1
********************** Run info ******************************
Blosc version: 1.5.3.dev ($Date:: 2014-12-30 #$)
Number of threads: 1
Using synthetic data with 19 significant bits (out of 32)
Working set: 512.0 MB
Dataset size: 536870912 bytes
Type size: 16 bytes
Element Count: 33554432
********************** Running benchmarks *********************
memcpy(write):		 47802.6 us, 10710.7 MB/s
memcpy(read):		 79542.9 us, 6436.8 MB/s
Compression level: 0
comp(write):	 58640.4 us, 8731.2 MB/s	  Final bytes: 536870928  Ratio: 1.00
decomp(read):	 50753.0 us, 10088.1 MB/s	  OK
Compression level: 1
comp(write):	 164283.9 us, 3116.6 MB/s	  Final bytes: 77791248  Ratio: 6.90
decomp(read):	 97233.7 us, 5265.7 MB/s	  OK
Compression level: 2
comp(write):	 178714.5 us, 2864.9 MB/s	  Final bytes: 45645840  Ratio: 11.76
decomp(read):	 107850.6 us, 4747.3 MB/s	  OK
Compression level: 3
comp(write):	 184140.8 us, 2780.5 MB/s	  Final bytes: 45645840  Ratio: 11.76
decomp(read):	 107683.2 us, 4754.7 MB/s	  OK
Compression level: 4
comp(write):	 226673.0 us, 2258.8 MB/s	  Final bytes: 40583184  Ratio: 13.23
decomp(read):	 107106.9 us, 4780.3 MB/s	  OK
Compression level: 5
comp(write):	 255675.8 us, 2002.5 MB/s	  Final bytes: 40583184  Ratio: 13.23
decomp(read):	 107019.9 us, 4784.2 MB/s	  OK
Compression level: 6
comp(write):	 281721.3 us, 1817.4 MB/s	  Final bytes: 23207952  Ratio: 23.13
decomp(read):	 127463.4 us, 4016.8 MB/s	  OK
Compression level: 7
comp(write):	 253457.5 us, 2020.1 MB/s	  Final bytes: 19865616  Ratio: 27.03
decomp(read):	 131386.4 us, 3896.9 MB/s	  OK
Compression level: 8
comp(write):	 254045.5 us, 2015.4 MB/s	  Final bytes: 19865616  Ratio: 27.03
decomp(read):	 132194.2 us, 3873.1 MB/s	  OK
Compression level: 9
comp(write):	 257042.1 us, 1991.9 MB/s	  Final bytes: 16814096  Ratio: 31.93
decomp(read):	 136268.1 us, 3757.3 MB/s	  OK

Round-trip compr/decompr on 15.0 GB
Elapsed time:	   11.1 s, 3031.7 MB/s
```