@FrancescAlted 2015-02-17
Looks pretty good, thanks!  I remember playing with unrolling too some years ago, but I don't remember seeing significant speed-ups.  Will do more tests in my machines and will report back.

Also, for some reason one of the tests in TravisCI is failing.