@FrancescAlted 2015-02-19
I just did some quick comparison and I did not find a significant improvement in my machine (i5-3380M CPU @ 2.90GHz, Ubuntu 14.10, gcc-4.9.1).  Here are my figures:

With master:

```
$ bench/bench blosclz single 1 | tee master1.out
Blosc version: 1.5.3.dev ($Date:: 2014-12-30 #$)
List of supported compressors in this build: blosclz,lz4,lz4hc,snappy,zlib
Supported compression libraries:
  BloscLZ: 1.0.3
  LZ4: 1.5.0
  Snappy: 1.1.1
  Zlib: 1.2.8
Using compressor: blosclz
Running suite: single
--> 1, 2097152, 8, 19, blosclz
********************** Run info ******************************
Blosc version: 1.5.3.dev ($Date:: 2014-12-30 #$)
Using synthetic data with 19 significant bits (out of 32)
Dataset size: 2097152 bytes     Type size: 8 bytes
Working set: 256.0 MB           Number of threads: 1
********************** Running benchmarks *********************
memcpy(write):            279.5 us, 7155.8 MB/s
memcpy(read):             231.5 us, 8638.0 MB/s
Compression level: 0
comp(write):      267.4 us, 7480.6 MB/s   Final bytes: 2097168  Ratio: 1.00
decomp(read):     249.5 us, 8017.0 MB/s   OK
Compression level: 1
comp(write):      860.7 us, 2323.8 MB/s   Final bytes: 1095184  Ratio: 1.91
decomp(read):     255.5 us, 7826.9 MB/s   OK
Compression level: 2
comp(write):      776.0 us, 2577.2 MB/s   Final bytes: 584976  Ratio: 3.59
decomp(read):     248.5 us, 8048.8 MB/s   OK
Compression level: 3
comp(write):      957.1 us, 2089.6 MB/s   Final bytes: 584976  Ratio: 3.59
decomp(read):     257.2 us, 7776.0 MB/s   OK
Compression level: 4
comp(write):     1443.0 us, 1386.0 MB/s   Final bytes: 558608  Ratio: 3.75
decomp(read):     250.7 us, 7977.2 MB/s   OK
Compression level: 5
comp(write):     1841.8 us, 1085.9 MB/s   Final bytes: 558608  Ratio: 3.75
decomp(read):     249.1 us, 8030.1 MB/s   OK
Compression level: 6
comp(write):     1823.6 us, 1096.8 MB/s   Final bytes: 348816  Ratio: 6.01
decomp(read):     842.1 us, 2375.1 MB/s   OK
Compression level: 7
comp(write):     1646.1 us, 1215.0 MB/s   Final bytes: 274768  Ratio: 7.63
decomp(read):     795.1 us, 2515.5 MB/s   OK
Compression level: 8
comp(write):     1638.6 us, 1220.6 MB/s   Final bytes: 274768  Ratio: 7.63
decomp(read):     795.5 us, 2514.0 MB/s   OK
Compression level: 9
comp(write):     1456.5 us, 1373.2 MB/s   Final bytes: 215872  Ratio: 9.71
decomp(read):     756.2 us, 2644.9 MB/s   OK

Round-trip compr/decompr on 7.5 GB
Elapsed time:       7.0 s, 2400.4 MB/s
```

with your unrolling work:

```
$ bench/bench blosclz single 1 | tee unrolling1.out
Blosc version: 1.5.3.dev ($Date:: 2014-12-30 #$)
List of supported compressors in this build: blosclz,lz4,lz4hc,snappy,zlib
Supported compression libraries:
  BloscLZ: 1.0.3
  LZ4: 1.5.0
  Snappy: 1.1.1
  Zlib: 1.2.8
Using compressor: blosclz
Running suite: single
--> 1, 2097152, 8, 19, blosclz
********************** Run info ******************************
Blosc version: 1.5.3.dev ($Date:: 2014-12-30 #$)
Using synthetic data with 19 significant bits (out of 32)
Dataset size: 2097152 bytes     Type size: 8 bytes
Working set: 256.0 MB           Number of threads: 1
********************** Running benchmarks *********************
memcpy(write):            280.8 us, 7121.3 MB/s
memcpy(read):             232.8 us, 8589.7 MB/s
Compression level: 0
comp(write):      291.7 us, 6855.9 MB/s   Final bytes: 2097168  Ratio: 1.00
decomp(read):     254.9 us, 7847.5 MB/s   OK
Compression level: 1
comp(write):      856.2 us, 2336.0 MB/s   Final bytes: 1095184  Ratio: 1.91
decomp(read):     254.0 us, 7872.5 MB/s   OK
Compression level: 2
comp(write):      773.2 us, 2586.6 MB/s   Final bytes: 584976  Ratio: 3.59
decomp(read):     257.8 us, 7759.1 MB/s   OK
Compression level: 3
comp(write):      862.2 us, 2319.6 MB/s   Final bytes: 584976  Ratio: 3.59
decomp(read):     249.5 us, 8014.4 MB/s   OK
Compression level: 4
comp(write):     1427.1 us, 1401.4 MB/s   Final bytes: 558608  Ratio: 3.75
decomp(read):     260.3 us, 7684.5 MB/s   OK
Compression level: 5
comp(write):     1810.1 us, 1104.9 MB/s   Final bytes: 558608  Ratio: 3.75
decomp(read):     250.1 us, 7995.8 MB/s   OK
Compression level: 6
comp(write):     1815.2 us, 1101.8 MB/s   Final bytes: 348816  Ratio: 6.01
decomp(read):     841.0 us, 2378.0 MB/s   OK
Compression level: 7
comp(write):     1614.2 us, 1239.0 MB/s   Final bytes: 274768  Ratio: 7.63
decomp(read):     789.6 us, 2532.9 MB/s   OK
Compression level: 8
comp(write):     1624.2 us, 1231.4 MB/s   Final bytes: 274768  Ratio: 7.63
decomp(read):     790.1 us, 2531.3 MB/s   OK
Compression level: 9
comp(write):     1454.3 us, 1375.2 MB/s   Final bytes: 215872  Ratio: 9.71
decomp(read):     749.8 us, 2667.2 MB/s   OK

Round-trip compr/decompr on 7.5 GB
Elapsed time:       7.0 s, 2424.7 MB/s
```

So provided that the unrolling definitely makes the code more obscure, I am not sold on this.  BTW, which is your benchmark environment (OS, C compiler version, ...)?