@jack-pappas 2015-02-22
@FrancescAlted I'm going to close this PR and open a separate one for the SSE2 shuffle/unshuffle support for large types, as it doesn't have anything to do with unrolling the loops.

For the loop unrolling, the benchmark environment I used was Windows 7 x64 and VS2013, and I used CMake to configure and build. I recompiled both the original and hand-unrolled versions with GCC 4.7 (from mingw) with ``-O3`` and ran the benchmarks again; this time, the results were basically the same between the two versions. So it appears that the version of Visual C++ which ships with VS2013 does not unroll the inner loops of the shuffle routines, which is why I saw the speedup after hand-unrolling them.

Out of curiosity, I also tried recompiling with the same GCC using ``-O2`` and ``-O2 -funroll-loops`` (``-O2`` does not include ``-funroll-loops``, while ``-O3`` does), and the performance improvement from specifying ``-funroll-loops`` looks similar to what I saw from hand-unrolling the loops and compiling with VC++.